refactor: #293 Intent 확장함수 생성
